+++ src/aux-common/runtime/AuxLibrary.spec.ts
@@ -2551,19 +2551,19 @@ describe('AuxLibrary', () => {
                 context.playerBot = player;
             });
 
-            it('should return true when pagePortal equals the given value', () => {
-                player.tags.pagePortal = 'dimension';
+            it('should return true when gridPortal equals the given value', () => {
+                player.tags.gridPortal = 'dimension';
                 const result = library.api.os.isInDimension('dimension');
                 expect(result).toEqual(true);
             });
 
-            it('should return false when pagePortal does not equal the given value', () => {
-                player.tags.pagePortal = 'dimension';
+            it('should return false when gridPortal does not equal the given value', () => {
+                player.tags.gridPortal = 'dimension';
                 const result = library.api.os.isInDimension('abc');
                 expect(result).toEqual(false);
             });
 
-            it('should return false when pagePortal is not set', () => {
+            it('should return false when gridPortal is not set', () => {
                 const result = library.api.os.isInDimension('dimension');
                 expect(result).toEqual(false);
             });
@@ -2571,7 +2571,7 @@ describe('AuxLibrary', () => {
             it.each(numberCases)(
                 'should support "%s" when given %s',
                 (expected, given) => {
-                    player.tags.pagePortal = given;
+                    player.tags.gridPortal = given;
                     const result = library.api.os.isInDimension(expected);
                     expect(result).toEqual(true);
                 }
@@ -2593,13 +2593,13 @@ describe('AuxLibrary', () => {
                 context.playerBot = player;
             });
 
-            it('should return pagePortal', () => {
-                player.tags.pagePortal = 'dimension';
+            it('should return gridPortal', () => {
+                player.tags.gridPortal = 'dimension';
                 const result = library.api.os.getCurrentDimension();
                 expect(result).toEqual('dimension');
             });
 
-            it('should return undefined when pagePortal is not set', () => {
+            it('should return undefined when gridPortal is not set', () => {
                 const result = library.api.os.getCurrentDimension();
                 expect(result).toBeUndefined();
             });
@@ -2607,7 +2607,7 @@ describe('AuxLibrary', () => {
             it.each(numberCases)(
                 'should return "%s" when given %s',
                 (expected, given) => {
-                    player.tags.pagePortal = given;
+                    player.tags.gridPortal = given;
                     const result = library.api.os.getCurrentDimension();
                     expect(result).toEqual(expected);
                 }
@@ -2683,8 +2683,8 @@ describe('AuxLibrary', () => {
                 context.playerBot = player;
             });
 
-            it('should return the miniPortal tag from the user bot', () => {
-                player.tags.miniPortal = 'abc';
+            it('should return the miniGridPortal tag from the user bot', () => {
+                player.tags.miniGridPortal = 'abc';
                 const result = library.api.os.getMiniPortalDimension();
                 expect(result).toEqual('abc');
             });
@@ -2692,7 +2692,7 @@ describe('AuxLibrary', () => {
             it.each(numberCases)(
                 'should return "%s" when given %s',
                 (expected, given) => {
-                    player.tags.miniPortal = given;
+                    player.tags.miniGridPortal = given;
                     const result = library.api.os.getMiniPortalDimension();
                     expect(result).toEqual(expected);
                 }
@@ -2734,12 +2734,12 @@ describe('AuxLibrary', () => {
             });
 
             const cases = [
-                ['page', 'pageDimension'],
-                ['pagePortal', 'pageDimension'],
+                ['grid', 'gridDimension'],
+                ['gridPortal', 'gridDimension'],
                 ['inventory', 'inventoryDimension'],
                 ['inventoryPortal', 'inventoryDimension'],
-                ['mini', 'miniDimension'],
-                ['miniPortal', 'miniDimension'],
+                ['miniGrid', 'miniDimension'],
+                ['miniGridPortal', 'miniDimension'],
                 ['menu', 'menuDimension'],
                 ['menuPortal', 'menuDimension'],
                 ['sheet', 'sheetDimension'],
@@ -2750,9 +2750,9 @@ describe('AuxLibrary', () => {
 
             describe.each(cases)('%s', (portal, expectedDimension) => {
                 it(`should get the dimension for the ${portal} portal`, () => {
-                    player.tags.pagePortal = 'pageDimension';
+                    player.tags.gridPortal = 'gridDimension';
                     player.tags.inventoryPortal = 'inventoryDimension';
-                    player.tags.miniPortal = 'miniDimension';
+                    player.tags.miniGridPortal = 'miniDimension';
                     player.tags.menuPortal = 'menuDimension';
                     player.tags.sheetPortal = 'sheetDimension';
                     player.tags.falsy = false;
@@ -2764,9 +2764,9 @@ describe('AuxLibrary', () => {
                 it.each(numberCases)(
                     'should return "%s" when given %s',
                     (expected, given) => {
-                        player.tags.pagePortal = given;
+                        player.tags.gridPortal = given;
                         player.tags.inventoryPortal = given;
-                        player.tags.miniPortal = given;
+                        player.tags.miniGridPortal = given;
                         player.tags.menuPortal = given;
                         player.tags.sheetPortal = given;
                         player.tags.falsy = false;
@@ -3020,30 +3020,30 @@ describe('AuxLibrary', () => {
                 context.playerBot = player;
             });
 
-            it('should return true if the given bot is in the users mini portal dimension', () => {
-                player.tags.miniPortal = 'abc';
+            it('should return true if the given bot is in the users miniGridPortal dimension', () => {
+                player.tags.miniGridPortal = 'abc';
                 bot1.tags.abc = true;
                 const result = library.api.os.hasBotInMiniPortal(bot1);
                 expect(result).toEqual(true);
             });
 
-            it('should return true if all the given bots are in the users mini portal dimension', () => {
-                player.tags.miniPortal = 'abc';
+            it('should return true if all the given bots are in the users miniGridPortal dimension', () => {
+                player.tags.miniGridPortal = 'abc';
                 bot1.tags.abc = true;
                 bot2.tags.abc = true;
                 const result = library.api.os.hasBotInMiniPortal([bot1, bot2]);
                 expect(result).toEqual(true);
             });
 
-            it('should return false if one of the given bots are not in the users mini portal dimension', () => {
-                player.tags.miniPortal = 'abc';
+            it('should return false if one of the given bots are not in the users miniGridPortal dimension', () => {
+                player.tags.miniGridPortal = 'abc';
                 bot1.tags.abc = false;
                 bot2.tags.abc = true;
                 const result = library.api.os.hasBotInMiniPortal([bot1, bot2]);
                 expect(result).toEqual(false);
             });
 
-            it('should return false if the player does not have an mini portal', () => {
+            it('should return false if the player does not have an miniGridPortal', () => {
                 bot1.tags.abc = true;
                 bot2.tags.abc = true;
                 const result = library.api.os.hasBotInMiniPortal([bot1, bot2]);
@@ -11103,12 +11103,12 @@ describe('AuxLibrary', () => {
     });
 
     describe('os.getCameraPosition()', () => {
-        let pagePortal: RuntimeBot;
-        let miniPortal: RuntimeBot;
+        let gridPortal: RuntimeBot;
+        let miniGridPortal: RuntimeBot;
 
         beforeEach(() => {
-            pagePortal = createDummyRuntimeBot(
-                'pagePortal',
+            gridPortal = createDummyRuntimeBot(
+                'gridPortal',
                 {
                     cameraPositionX: 1,
                     cameraPositionY: 2,
@@ -11116,8 +11116,8 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            miniPortal = createDummyRuntimeBot(
-                'miniPortal',
+            miniGridPortal = createDummyRuntimeBot(
+                'miniGridPortal',
                 {
                     cameraPositionX: 4,
                     cameraPositionY: 5,
@@ -11125,19 +11125,19 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            addToContext(context, pagePortal, miniPortal);
+            addToContext(context, gridPortal, miniGridPortal);
 
-            (<any>globalThis).pagePortalBot = pagePortal;
-            (<any>globalThis).miniPortalBot = miniPortal;
+            (<any>globalThis).gridPortalBot = gridPortal;
+            (<any>globalThis).miniGridPortalBot = miniGridPortal;
         });
 
         afterEach(() => {
-            delete (<any>globalThis).pagePortalBot;
-            delete (<any>globalThis).miniPortalBot;
+            delete (<any>globalThis).gridPortalBot;
+            delete (<any>globalThis).miniGridPortalBot;
         });
 
-        it('should return NaN for x, y, and z if the page portal bot is null', () => {
-            (<any>globalThis).pagePortalBot = null;
+        it('should return NaN for x, y, and z if the grid portal bot is null', () => {
+            (<any>globalThis).gridPortalBot = null;
             const result = library.api.os.getCameraPosition();
 
             expect(result).toEqual({
@@ -11147,7 +11147,7 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should return the x, y, and z of the camera for the page portal', () => {
+        it('should return the x, y, and z of the camera for the grid portal', () => {
             const result = library.api.os.getCameraPosition();
 
             expect(result).toEqual({
@@ -11158,7 +11158,7 @@ describe('AuxLibrary', () => {
         });
 
         it('should be able to get the mini camera position', () => {
-            const result = library.api.os.getCameraPosition('mini');
+            const result = library.api.os.getCameraPosition('miniGrid');
 
             expect(result).toEqual({
                 x: 4,
@@ -11167,8 +11167,8 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should be able to get the page camera position', () => {
-            const result = library.api.os.getCameraPosition('page');
+        it('should be able to get the bot camera position', () => {
+            const result = library.api.os.getCameraPosition('grid');
 
             expect(result).toEqual({
                 x: 1,
@@ -11179,12 +11179,12 @@ describe('AuxLibrary', () => {
     });
 
     describe('os.getCameraRotation()', () => {
-        let pagePortal: RuntimeBot;
-        let miniPortal: RuntimeBot;
+        let gridPortal: RuntimeBot;
+        let miniGridPortal: RuntimeBot;
 
         beforeEach(() => {
-            pagePortal = createDummyRuntimeBot(
-                'pagePortal',
+            gridPortal = createDummyRuntimeBot(
+                'gridPortal',
                 {
                     cameraRotationX: 1,
                     cameraRotationY: 2,
@@ -11192,8 +11192,8 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            miniPortal = createDummyRuntimeBot(
-                'miniPortal',
+            miniGridPortal = createDummyRuntimeBot(
+                'miniGridPortal',
                 {
                     cameraRotationX: 4,
                     cameraRotationY: 5,
@@ -11201,19 +11201,19 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            addToContext(context, pagePortal, miniPortal);
+            addToContext(context, gridPortal, miniGridPortal);
 
-            (<any>globalThis).pagePortalBot = pagePortal;
-            (<any>globalThis).miniPortalBot = miniPortal;
+            (<any>globalThis).gridPortalBot = gridPortal;
+            (<any>globalThis).miniGridPortalBot = miniGridPortal;
         });
 
         afterEach(() => {
-            delete (<any>globalThis).pagePortalBot;
-            delete (<any>globalThis).miniPortalBot;
+            delete (<any>globalThis).gridPortalBot;
+            delete (<any>globalThis).miniGridPortalBot;
         });
 
-        it('should return NaN for x, y, and z if the page portal bot is null', () => {
-            delete (<any>globalThis).pagePortalBot;
+        it('should return NaN for x, y, and z if the grid portal bot is null', () => {
+            delete (<any>globalThis).gridPortalBot;
             const result = library.api.os.getCameraRotation();
 
             expect(result).toEqual({
@@ -11223,7 +11223,7 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should return the x, y, and z of the player camera for the page portal', () => {
+        it('should return the x, y, and z of the player camera for the grid portal', () => {
             const result = library.api.os.getCameraRotation();
 
             expect(result).toEqual({
@@ -11233,8 +11233,8 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should be able to get the mini portal camera rotation', () => {
-            const result = library.api.os.getCameraRotation('mini');
+        it('should be able to get the miniGridPortal camera rotation', () => {
+            const result = library.api.os.getCameraRotation('miniGrid');
 
             expect(result).toEqual({
                 x: 4,
@@ -11243,8 +11243,8 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should be able to get the page camera rotation', () => {
-            const result = library.api.os.getCameraRotation('page');
+        it('should be able to get the grid camera rotation', () => {
+            const result = library.api.os.getCameraRotation('grid');
 
             expect(result).toEqual({
                 x: 1,
@@ -11255,12 +11255,12 @@ describe('AuxLibrary', () => {
     });
 
     describe('os.getFocusPoint()', () => {
-        let pagePortal: RuntimeBot;
-        let miniPortal: RuntimeBot;
+        let gridPortal: RuntimeBot;
+        let miniGridPortal: RuntimeBot;
 
         beforeEach(() => {
-            pagePortal = createDummyRuntimeBot(
-                'pagePortal',
+            gridPortal = createDummyRuntimeBot(
+                'gridPortal',
                 {
                     cameraFocusX: 1,
                     cameraFocusY: 2,
@@ -11268,8 +11268,8 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            miniPortal = createDummyRuntimeBot(
-                'miniPortal',
+            miniGridPortal = createDummyRuntimeBot(
+                'miniGridPortal',
                 {
                     cameraFocusX: 4,
                     cameraFocusY: 5,
@@ -11277,19 +11277,19 @@ describe('AuxLibrary', () => {
                 },
                 'tempLocal'
             );
-            addToContext(context, pagePortal, miniPortal);
+            addToContext(context, gridPortal, miniGridPortal);
 
-            (<any>globalThis).pagePortalBot = pagePortal;
-            (<any>globalThis).miniPortalBot = miniPortal;
+            (<any>globalThis).gridPortalBot = gridPortal;
+            (<any>globalThis).miniGridPortalBot = miniGridPortal;
         });
 
         afterEach(() => {
-            delete (<any>globalThis).pagePortalBot;
-            delete (<any>globalThis).miniPortalBot;
+            delete (<any>globalThis).gridPortalBot;
+            delete (<any>globalThis).miniGridPortalBot;
         });
 
-        it('should return NaN for x, y, and z if the page portal bot is null', () => {
-            delete (<any>globalThis).pagePortalBot;
+        it('should return NaN for x, y, and z if the grid portal bot is null', () => {
+            delete (<any>globalThis).gridPortalBot;
             const result = library.api.os.getFocusPoint();
 
             expect(result).toEqual({
@@ -11299,7 +11299,7 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should return the x, y, and z of the player camera for the page portal', () => {
+        it('should return the x, y, and z of the player camera for the grid portal', () => {
             const result = library.api.os.getFocusPoint();
 
             expect(result).toEqual({
@@ -11309,8 +11309,8 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should be able to get the mini portal camera rotation', () => {
-            const result = library.api.os.getFocusPoint('mini');
+        it('should be able to get the miniGridPortal camera rotation', () => {
+            const result = library.api.os.getFocusPoint('miniGrid');
 
             expect(result).toEqual({
                 x: 4,
@@ -11319,8 +11319,8 @@ describe('AuxLibrary', () => {
             });
         });
 
-        it('should be able to get the page camera rotation', () => {
-            const result = library.api.os.getFocusPoint('page');
+        it('should be able to get the grid camera rotation', () => {
+            const result = library.api.os.getFocusPoint('grid');
 
             expect(result).toEqual({
                 x: 1,