+++ src/aux-server/aux-web/aux-player/PlayerHome/PlayerHome.ts
@@ -112,30 +112,40 @@ export default class PlayerHome extends Vue {
         if (this.query) {
             // On first load check the inst and load a default
             let inst = this.query['inst'] as string | string[];
+            let update: Dictionary<string | string[]> = {};
             if (!hasValue(inst)) {
                 // if there is no inst tag defined, check for the story tag and then the server tag
                 inst = this.query['story'] ?? this.query['server'];
                 if (hasValue(inst)) {
-                    let update: Dictionary<string | string[]> = {
-                        inst: inst,
-                        story: null,
-                    };
-                    this._updateQuery(update);
+                    update.inst = inst;
+                    update.story = null;
+                    update.server = null;
                 } else {
                     // Generate a random inst name
                     const randomName: string = uniqueNamesGenerator(
                         namesConfig
                     );
-                    let update: Dictionary<string> = {};
                     if (!appManager.config.disableCollaboration) {
                         update.inst = randomName;
                     }
-                    if (!hasValue(this.query['pagePortal'])) {
-                        update.pagePortal = 'home';
+                    if (!hasValue(this.query['gridPortal'])) {
+                        update.gridPortal = 'home';
                     }
-                    this._updateQuery(update);
                     inst = randomName;
                 }
+            }
+
+            if (
+                hasValue(this.query['pagePortal']) &&
+                !hasValue(this.query['gridPortal'])
+            ) {
+                const portal = this.query['pagePortal'];
+                update.pagePortal = null;
+                update.gridPortal = Array.isArray(portal) ? portal[0] : portal;
+            }
+
+            if (Object.keys(update).length > 0) {
+                this._updateQuery(update);
             }
             this._setServer(inst);
         }